# kapsule-incus-manager/api/schema/openapi.yaml
openapi: "3.1.0"
info:
  title: Kapsule-Incus-Manager API
  version: "0.1.0"
  description: >
    Canonical API definition for the KIM daemon.
    All three frontends (QML, web, CLI) are thin clients against this API.
    The D-Bus interface mirrors every operation defined here.
  license:
    name: GPL-3.0-or-later
servers:
  - url: http://127.0.0.1:8765
    description: Local KIM daemon
...
components:
  schemas:
...
    Operation:
      type: object
      required: [id, description, status, created_at]
      properties:
        id:          { type: string, format: uuid }
        description: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
        metadata:    { type: object, additionalProperties: true }
        error:       { type: string }
    InstanceStatus:
      type: string
      enum: [Running, Stopped, Frozen, Error, Unknown]
    InstanceType:
      type: string
      enum: [container, virtual-machine]
    InstanceSummary:
      type: object
      required: [name, type, status, project, remote]
      properties:
        name:       { type: string }
        type:       { $ref: "#/components/schemas/InstanceType" }
        status:     { $ref: "#/components/schemas/InstanceStatus" }
        image:      { type: string }
        project:    { type: string }
        remote:     { type: string }
        created_at: { type: string, format: date-time }
        cpu_usage:  { type: number, description: "CPU usage 0.0–1.0" }
        memory_usage_bytes: { type: integer }
        disk_usage_bytes:   { type: integer }
...
  responses:
...
    Accepted:
      description: Operation accepted and running asynchronously
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Operation" }
...
paths:
...
  /api/v1/provisioning/windows:
    post:
      operationId: createWindowsVm
      summary: Create a Windows VM in Incus
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:          { type: string, default: "windows" }
                profile:       { type: string, default: "windows-desktop" }
                image:         { type: string, description: "Path to Windows ISO" }
                disk:          { type: string, description: "Path to additional disk image" }
                ram:           { type: string, default: "8GiB" }
                cpus:          { type: integer, default: 4 }
                disk_size:     { type: string, default: "64GB" }
                gpu_overlay:   { type: string, description: "GPU overlay profile (vfio, looking-glass)" }
                rdp:           { type: boolean, default: true }
                boot_autostart: { type: boolean, default: false }
                project:       { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/start:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: startWindowsVm
      summary: Start a Windows VM
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/stop:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: force, in: query, schema: { type: boolean, default: false } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: stopWindowsVm
      summary: Stop a Windows VM
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/snapshots:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listWindowsSnapshots
      summary: List snapshots for a Windows VM
      tags: [provisioning/windows]
      responses:
        "200":
          description: Snapshot list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
    post:
      operationId: createWindowsSnapshot
      summary: Create a snapshot
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot]
              properties:
                snapshot: { type: string }
                stateful: { type: boolean, default: false }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/snapshots/schedule:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      operationId: setWindowsSnapshotSchedule
      summary: Set automatic snapshot schedule for a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule]
              properties:
                schedule: { type: string }
                expiry:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "200":
          description: Schedule updated
  /api/v1/provisioning/windows/{name}/snapshots/{snapshot}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: restoreWindowsSnapshot
      summary: Restore a Windows VM snapshot
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/snapshots/{snapshot}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: deleteWindowsSnapshot
      summary: Delete a Windows VM snapshot
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/backups:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listWindowsBackups
      summary: List Incus backups for a Windows VM
      tags: [provisioning/windows]
      responses:
        "200":
          description: Backup list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
    post:
      operationId: backupWindowsVm
      summary: Create an Incus backup of a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backup_name:       { type: string }
                instance_only:     { type: boolean, default: false }
                optimized_storage: { type: boolean, default: true }
                project:           { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: restoreWindowsVmBackup
      summary: Restore a Windows VM from an Incus backup
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [backup_name]
              properties:
                backup_name: { type: string }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/gpus:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: attachWindowsGpu
      summary: Attach a GPU to a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dev_name: { type: string, default: "gpu0" }
                gpu_type: { type: string, default: physical }
                pci:      { type: string }
                vendor:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/gpus/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: detachWindowsGpu
      summary: Detach a GPU from a Windows VM
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/forwards:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: addWindowsForward
      summary: Add a port-forward proxy device to a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_port, guest_port]
              properties:
                host_port:   { type: integer }
                guest_port:  { type: integer }
                protocol:    { type: string, default: "tcp" }
                listen_addr: { type: string, default: "127.0.0.1" }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/forwards/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: removeWindowsForward
      summary: Remove a port-forward proxy device from a Windows VM
      tags: [provisioning/windows]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/guest-tools:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: installWindowsGuestTools
      summary: Install Windows guest tools via PowerShell exec
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tools:   { type: array, items: { type: string, enum: [svcguest, serviceman, srvlib, winbtrfs, winfsp] } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/remoteapp:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: discoverWindowsRemoteApps
      summary: Discover installed Windows applications
      tags: [provisioning/windows]
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/v1/provisioning/windows/{name}/remoteapp/launch:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: launchWindowsRemoteApp
      summary: Launch a Windows application inside the VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app]
              properties:
                app:     { type: string }
                args:    { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/apps:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: installWindowsApps
      summary: Install Windows applications via winget
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apps]
              properties:
                apps:    { type: array, items: { type: string }, description: "winget package IDs" }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/cloud-sync:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      operationId: configureWindowsCloudSync
      summary: Configure rclone cloud sync for Windows VM backups
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [remote_name, remote_path]
              properties:
                remote_name: { type: string }
                remote_path: { type: string }
                schedule:    { type: string }
                project:     { type: string, default: "" }
      responses:
        "200":
          description: Cloud sync configured
  /api/v1/provisioning/windows/{name}/harden:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: hardenWindowsVm
      summary: Apply security hardening to a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:   { type: string, enum: [basic, standard, strict], default: standard }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/{name}/disk/resize:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: resizeWindowsDisk
      summary: Resize the root disk of a Windows VM
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_size]
              properties:
                new_size: { type: string, example: "128GB" }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/fleet:
    get:
      operationId: listWindowsFleet
      summary: List Windows VMs
      tags: [provisioning/windows]
      parameters:
        - { name: project, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: VM list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InstanceSummary" } }
  /api/v1/provisioning/windows/fleet/start:
    post:
      operationId: fleetStartWindows
      summary: Start multiple Windows VMs
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/fleet/stop:
    post:
      operationId: fleetStopWindows
      summary: Stop multiple Windows VMs
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/windows/publish:
    post:
      operationId: publishWindowsVm
      summary: Publish a Windows VM as a reusable Incus image
      tags: [provisioning/windows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                alias:       { type: string }
                description: { type: string }
                public:      { type: boolean, default: false }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }